requirements.txt
python-telegram-bot==13.8.1
google-cloud-dialogflow==2.9.1
vk-api==11.9.6
redis==4.0.0
hiredis==2.0.0